+++ src/layouts/InterfaceLayout/containers/SwapContainer/components/ProvidersRadioSelector/ProvidersRadioSelector.vue
@@ -22,7 +22,7 @@
             <label :for="provider.provider" />
           </div>
           <div class="provider-image">
-            <img :src="providerLogo(provider.provider)" />
+            <img :src="providerLogo(provider)" />
           </div>
           <div
             :class="[
@@ -171,6 +171,7 @@ import KyberNetwork from '@/assets/images/etc/kybernetowrk.png';
 import Bity from '@/assets/images/etc/bity.png';
 import Simplex from '@/assets/images/etc/simplex.png';
 import Changelly from '@/assets/images/etc/changelly.png';
+import bityBeta from '@/assets/images/etc/bitybeta.png';
 
 export default {
   props: {
@@ -231,6 +232,9 @@ export default {
         bity: Bity,
         simplex: Simplex,
         changelly: Changelly
+      },
+      betaLogos: {
+        bity: bityBeta
       }
     };
   },
@@ -258,8 +262,15 @@ export default {
       clickedEl.classList.add('radio-selected');
       this.$emit('selectedProvider', provider);
     },
-    providerLogo(name) {
-      return this.logos[name];
+    providerLogo(details) {
+      if (this.useBetaLogo(details)) return this.betaLogos[details.provider];
+      return this.logos[details.provider];
+    },
+    useBetaLogo(details) {
+      return (
+        details.provider === 'bity' &&
+        (details.toCurrency === 'EUR' || details.toCurrency === 'CHF')
+      );
     },
     minNote(details) {
       if (details.minValue > 0) {